+++ src/settings.js
@@ -592,6 +592,7 @@ var ASYNCIFY_STACK_SIZE = 4096;
 // changes which would mean a single list couldn't work for both -O0 and -O1
 // builds, etc.). You can inspect the wasm binary to look for the actual names,
 // either directly or using wasm-objdump or wasm-dis, etc.
+// Simple '*' wildcard matching is supported.
 var ASYNCIFY_BLACKLIST = [];
 
 // If the Asyncify whitelist is provided, then *only* the functions in the list
@@ -1401,7 +1402,8 @@ var FETCH = 0;
 // only relevant for fastcomp, where we support asm.js. As a result, some
 // synchronous fetch operations that depend on the fetch worker may not work
 // with the wasm backend, like waiting or IndexedDB.
-var USE_FETCH_WORKER = 0;
+// Currently will always be set to 0 on WASM backend.
+var USE_FETCH_WORKER = 1;
 
 // Internal: name of the file containing the Fetch *.fetch.js, if relevant
 // Do not set yourself.